+++ src/dpr/routes/journeys/download-report/sync-tests.cy.ts
@@ -98,17 +98,17 @@ context('Platform sync download tests', () => {
       )
       cy.findByRole('link', { name: /Return to report to download/ }).click()
       cy.findAllByRole('heading').contains('Sync report').should('exist')
-      cy.findByRole('button', { name: /download/ }).should('be.visible')
+      cy.findByRole('button', { name: /Download/ }).should('be.visible')
 
       cy.task('stubSyncReportDownload')
-      cy.findByRole('button', { name: /download/ }).click()
+      cy.findByRole('button', { name: /Download/ }).click()
       cy.task('checkCsvDownload4RowsValid').should('equal', true)
     })
 
     it('should redirect on trying to download after having the permission to download removed', () => {
       updateRedisState('downloadPermissions', [])
       cy.findByRole('heading', { name: /To download this report/ }).should('not.exist')
-      cy.findByRole('button', { name: /download/ }).click()
+      cy.findByRole('button', { name: /Enable download/ }).click()
       cy.findByRole('heading', { name: /To download this report/ }).should('be.visible')
     })
   })