+++ test-app/routes/components/report-heading/routes.ts
@@ -1,0 +1,10 @@
+/* eslint-disable no-param-reassign */
+import { Router } from 'express'
+import { ReportHeadingController } from './controller'
+
+export function ReportHeadingRoutes(): Router {
+  const router = Router({ mergeParams: true })
+  const controller = new ReportHeadingController()
+  router.get('/', controller.GET)
+  return router
+}